Rebase and doc update
