+++ docs/source/generating-a-node.rst
@@ -97,7 +97,12 @@ nodes. Here is an example ``Cordform`` task called ``deployNodes`` that creates 
             h2Port   10008
             cordapps = ["net.corda:corda-finance:$corda_release_version"]
             // Grants user1 all RPC permissions.
-            rpcUsers = [[ user: "user1", "password": "test", "permissions": ["ALL"]]]
+            rpcUsers = users {
+                "user1" {
+                    password = "test"
+                    permissions = ["ALL"]
+                }
+            }
         }
         node {
             name "O=PartyB,L=New York,C=US"
@@ -107,7 +112,12 @@ nodes. Here is an example ``Cordform`` task called ``deployNodes`` that creates 
             h2Port   10012
             cordapps = ["net.corda:corda-finance:$corda_release_version"]
             // Grants user1 the ability to start the MyFlow flow.
-            rpcUsers = [[ user: "user1", "password": "test", "permissions": ["StartFlow.net.corda.flows.MyFlow"]]]
+            rpcUsers = users {
+               "user1" {
+                   password = "test"
+                   permissions = ["StartFlow.net.corda.flows.MyFlow"]
+               }
+            }
         }
     }
